Introduces zipWithIndex by row id
This commit add the ability to introduce a gobally sorted row index to
the inputs of a file. In contrast to the normal zipWithIndex from the
Flink utils, it not only guarantees a global ordering, but a global
ordering where the index reflects the actual row/line number of the
input file.